# fantastic-coffee-decaffeinated/doc/api.yaml
openapi: 3.0.1
info:
  title: WASAPhoto
  description: |
    This API allows to manage an instagram-like web application
  version: 1.0.0
servers:
  - url: http://localhost:3000
...
paths:
...
  /my-stream:
    get:
      tags: ["user"]
      operationId: getMyStream
      summary: Return the logged user stream
      description: |
        Get the user stream consisting in a list of thumbnailObjects
      responses:
        "200":
          description: Downloaded the user stream
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/thumbnailObject"
                minItems: 0
                maxItems: 1000
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          description: The server cannot elaborate the getMyStream request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
  /posts/{idPhoto}:
    parameters:
      - $ref: "#/components/parameters/idPhoto"
    get:
      tags: ["photo"]
      operationId: getPost
      summary: get a user post
      description: |
        Get a full user post
        with the photo, likes number, date time and comments list
      responses:
        "200":
          description: post downloaded
          content:           
            application/json:
              schema:
                $ref: "#/components/schemas/postObject"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "404":
          description: Post not found for get operation
          content:
           application/json:
            schema:
              $ref: "#/components/schemas/generalError"
        '500':
          description: The server cannot elaborate the getPost request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
  /photos/{idPhoto}:
    parameters:
      - $ref: "#/components/parameters/idPhoto"
    get:
      tags: ["photo"]
      operationId: getPhoto
      summary: get a user photo
      description: |
        Get a full user photo (not thumbnail)
        with likes number and comments list
      responses:
        "200":
          description: Photo, likes and comments downloaded
          content:       
            image/*:     
             schema: 
               type: string
               format: binary
        "404":
          description: Photo not found for getPhoto operation
          content:
           application/json:
            schema:
              $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        '500':
          description: The server cannot elaborate the getPhoto request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
    delete:
      tags: ["photo"]
      operationId: deletePhoto
      summary: Delete a user photo
      description: |
        Delete a photo from the current user logged in
      responses:
        "200":
          description: Photo deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalFeedback"
        "400":
          description: Invalid user id or photo id for deletePhoto operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "404":
          description: Photo not found for delete operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /follows:
    post:
      tags: ["user"]
      operationId: followUser
      summary: Follow one user
      description: |
        Allow to follow the username in the path
      requestBody:
        content:
          application/json:
            schema:
              description: the username of the user to follow
              type: object
              properties:
                name:
                  description: The user's username to follow
                  type: string
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
            examples:
              user:
                summary: maria-user
                value:
                  name: maria
        required: true
      responses:
        "201":
          description: User followed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalFeedback"
        "400":
          description: Invalid user id for followUser operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "500":
          description: The server cannot proces the followUser request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
...
  /photos/{idPhoto}/likes:
    parameters:
      - $ref: "#/components/parameters/idPhoto"
    post: 
      tags: ["photo"]
      operationId: likePhoto
      summary: Give a like to the photo identified by idPhoto
      description: |
        Allow to post a like on a user photo, so username is the visited user and
        idPhoto is the photo to like
      responses:
        "201":
          description: Like done on likePhoto operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalFeedback"
        "400":
          description: Photo format/photo id/user id not valid for likePhoto operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
        "500":
          description: The server cannot proces the likePhoto request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
  /photos/{idPhoto}/likes/{username}:
    parameters:
      - $ref: "#/components/parameters/idPhoto"
      - $ref: "#/components/parameters/username"  
    delete:
      tags: ["photo"]
      operationId: removeLike  
      summary: remove a like to a photo
      description: |
        Remove a like to the photo identified by idPhoto
      responses:
        "200":
          description: like removed on removeLike operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalFeedback"
        "400":
          description: Photo format/photo id/user id not valid for unlikePhoto operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
  /photos/{idPhoto}/comments:
    parameters:
      - $ref: "#/components/parameters/idPhoto"
    post: 
      tags: ["photo"]
      operationId: commentPhoto
      summary: Give a comment to a photo
      description: |
        Allow to post a comment on a user photo
      requestBody:
        content:
          application/json:
            schema:
              description: |
                The comment to post to a photo
              type: object
              properties:
                comment:
                  description: The content of the comment
                  type: string
            examples:
              photocomment:
                summary: the maria comment to the idPhoto photo
                value:
                  comment: this photo is really bad!
      responses:
        "201":
          description: comment done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalFeedback"
        "400":
          description: |
            Photo format/photo id/user id/comment length not valid for commentPhoto operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
    get: 
      tags: ["photo"]
      operationId: getComments
      summary: Get the comments for a photo
      description: |
        Allow to get the comments posted on a user photo
      responses:
        "201":
          description: the returned comments list
          content:
            application/json:
              schema:
                description: |
                  a comments object listing all the comments related to a photo
                type: object
                properties:
                  comments:
                    description: a single comment related to the photo
                    type: array
                    items:
                      $ref: "#/components/schemas/comment"
                    minItems: 0
                    maxItems: 1000
        "400":
          description: comment format/photo id/user id not valid for getComments operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
  /bans:
    post: 
      tags: ["user"]
      operationId: banUser
      summary: ban one user
      description: |
        Allow to ban a user
      requestBody:
        content:
          application/json:
            schema:
              description: the banned user to insert in the bans list
              type: object
              properties:
                name:
                  description: The banned user's username
                  type: string
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
            examples:
              user:
                summary: maria-user
                value:
                  name: maria
        required: true
      responses:
        "201":
          description: ban done
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalFeedback"
        "400":
          description: user id not valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalError"
        "401":
          $ref: "#/components/responses/UnauthorizedError"
...
components:
  securitySchemes:
    bearerAuth:   
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    username:
      name: username
      in: path
      required: true
      description: A username
      schema:
        type: string
    idPhoto:
      name: idPhoto
      in: path
      required: true
      description: a photo Id to identify a photo
      schema:
        type: string
...
  schemas:
    thumbnailObject:
      description: |
        A rappresentation of a thumnail pic with other informations
      type: object
      properties:
        idphoto:
          description: the photo url
          type: string
        datetime:
          description: the post date time to the referred photo
          type: string
          format: date-time
        nlikes:
          description: The likes number to the referred photo
          type: integer
        ncomments:
          description: The comments number to the referred photo
          type: integer
      example:
        photo1:
          idphoto: idphoto1234
          datetime: "2022-02-02T13:30:15"
          nlikes: 10
          ncomments: 3
        photo2:
          idphoto: idphoto4321
          datetime: "2022-03-09T17:20:24"
          nlikes: 5
          ncomments: 2
    postObject:
      description: |
        An object with the original photo (not thumbnail), likes number
        and comment list
      type: object
      properties:
        originalImage:
          description: The image url
          type: string
        nLikes:
          description: The likes number referred to the photo
          type: integer
        comments:
          description: The comments list referred to the photo
          type: array
          items:
            $ref: "#/components/schemas/comment"
          minItems: 0
          maxItems: 1000
    comment:
      description: |
        A comment by one user to one image, it contains che username of the
        comment owner and the comment itself
      type: object
      properties:
        username:
          description: The username of the comment
          type: string
        comment:
          description: The comment content
          type: string
      example:
        comment1:
          username: "pippo"
          comment: "wow!"
        comment2:
          username: "pluto"
          comment: "woof!"
    generalError:
      description: An object describing an error
      type: object
      properties:
        error:
          description: The message content of the error
          type: string
          example: "error description"
    generalFeedback:
      description: The feedback provided from a response
      type: object
      properties:
        feedback:
          description: The content of the feedback
          type: string
          example: "feedback description"
  responses:
    UnauthorizedError:
      description: Access identifier is missing or invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/generalError"
security:
 - bearerAuth: []